Work around MBEDTLS_NET_C being undefined in hxcpp
MBEDTLS_NET_C has historically been undefined in the hxcpp config for
mbedtls. This is an issue for curl because it depends on functions
defined in net_sockets.c. To work around this, we can create a
MBEDTLS_USER_CONFIG_FILE to ensure that MBEDTLS_NET_C is set.

On Windows, in order for net_sockets.c to compile properly with the
mbedtls threading configuration, we need to also set
`WIN32_LEAN_AND_MEAN`.

Some versions of hxcpp do have a MBEDTLS_USER_CONFIG_FILE already set,
in this case we may get a warning about overwriting the flag.