Fix static build with mingw
`wbemuuid.lib` must be linked explicitly, since #pragma comment(lib, ...)
only works with msvc.

`comsuppw.lib` does not exist in mingw.

`no_shared_libs` must be set for mingw to ensure the standard library
and other libraries are linked statically.

`.lib` must be replaced with `${LIBEXT}` so that it resolves to `.a` on
mingw.